requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
pytesseract
pillow
openai
endee
